SLEMA-97 Split types in assessment.dart into multiple files
