+++ SlemaForAndroid/lib/features/well_being/logic/entity/enum/assessment_factory_impl.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import 'package:pg_slema/features/well_being/logic/entity/assessment.dart';
 import 'package:pg_slema/features/well_being/logic/entity/assessment_factory.dart';
+import 'package:pg_slema/features/well_being/logic/entity/enum/symptom_entries.dart';
 import 'package:pg_slema/features/well_being/logic/entity/enum/well_being_variant.dart';
 
 class AssessmentFactoryImpl implements AssessmentFactory {
